ChangedLocationOfRevenuePerMovie
i changed the location of when u go to Admin>FilmOverzicht the TotaleOmzetPerFilm to Admin>Omzet>TotaleOmzetPerFilm. Admin>Omzet now has : TotaleOmzet,TotaleOmzetPerFilm and Back function. FilmOverzicht is now FilmPlannen. You can only plan a movie in FilmPlannen there so i removed the option of choosing to schedule the film or see the revenue of the film